affichage KO mais structure plus claire (?)

+++ src/Texture.cpp
@@ -29,7 +29,7 @@ void Texture::load_PNG(const char *filename)
 
     texture = new Color[width * height];
 
-    for (size_t i = 0; i < width * height; i++)
+    for (int i = 0; i < width * height; i++)
         texture[i] = _texture[i];
     
     stbi_image_free(_texture);
